requirements.txt
google-generativeai==0.4.1
google-ai-generativelanguage==0.4.0
pinecone-client>=5.0.0,<6.0.0
fastapi==0.109.1
uvicorn==0.27.1
gunicorn==21.2.0
python-dotenv==1.0.0
pydantic<2
langchain==0.1.16
langchain-core==0.1.52
langchain-community==0.0.34
sentence-transformers==2.6.0
huggingface-hub==0.22.2
playwright==1.42.0
beautifulsoup4==4.12.3
typing-extensions>=4.8.0
numpy>=1.26.4
requests>=2.32.3
psutil